# Treat unencoded `/` inside quoted literals as data (not separators)