<!-- markdownlint-disable MD002 MD041 -->

## Issues

*This pull request fixes https://github.com/OData/AspNetCoreOData/issues/709, fixes https://github.com/OData/odata.net/issues/2292, fixes https://github.com/OData/WebApi/issues/2309, fixes https://github.com/OData/WebApi/issues/1964, fixes https://github.com/OData/odata.net/issues/2477, fixes https://github.com/OData/odata.net/issues/735, fixes https://github.com/OData/odata.net/issues/1708, fixes https://github.com/OData/odata.net/issues/3327

## Description

This change updates `ParsePathIntoSegments` to:

- Parse the escaped path and split only on unquoted `/`, preserving any unencoded slashes inside quoted literals as part of the same segment for compatibility with some external clients - **reported case referred to Excel as one such integration https://github.com/OData/AspNetCoreOData/issues/709**.  
- Keep the existing behavior for properly **percent‑encoded** slashes (`%2F`) and for balanced quotes (including doubled single quotes for escaping).  
- Preserve prior error behavior for malformed percent escapes and unbalanced quotes.  

This aligns segment parsing with the [OData URL Conventions](http://docs.oasis-open.org/odata/odata/v4.0/errata03/os/complete/part2-url-conventions/odata-v4.0-errata03-os-part2-url-conventions-complete.html#_Toc453752335) (slashes that are part of a literal/key **must** be percent‑encoded) and avoids producing OData‑inapplicable segment shapes when a client violates that rule. 

### Motivation & Standards Alignment

- **OData 4.01 URL Conventions** require that forward slashes in key value segments be percent‑encoded. If a client sends `/People('foo/bar')`, the slash inside the literal should have been `%2F`. When this is not respected, treating the raw `/` as a separator results in segments that **do not correspond to any valid OData path tokenization**. This PR handles that non‑conformance more gracefully by preserving the literal as a single segment.   
- **RFC 3986** defines `/` as a path segment delimiter. Parsing should use the escaped path so that `%2F` stays data while raw `/` remains a separator. This change does exactly that: we operate on the escaped path and only treat unquoted `/` as a separator. 

### Behavior change: before vs after

#### Before (current behavior with non‑conformant clients)
- Request:  
  `GET /Service.svc/People('foo/bar')` *(slash not encoded within the literal)*
- Parsed segments:  
  `["People('foo", "bar')"]` <-- Two segments that are not meaningful OData segments.

#### After (this PR)
- Same request:  
  `GET /Service.svc/People('foo/bar')`
- Parsed segments:  
  `["People('foo/bar')"]`  <-- Single segment, quote‑consistent and usable for OData path handling.

#### No change for conformant inputs
- Request:  
  `GET /Service.svc/People('foo%2Fbar')`
- Parsed segments (before and after):  
  `["People('foo%2Fbar')"]`  <-- slash remains within the literal; downstream can unescape as needed.  
  *(OData 4.01 explicitly requires this encoding.)* 

### Is this a breaking change?

**Technically, yes**  - code that intentionally depended on the previously split result (`["People('foo", "bar')"]`) would now see a single segment. However:

- That two‑segment result was an inapplicable outcome for OData: neither segment is a valid OData path segment on its own; routing, model binding, and `ODataUriParser` semantics do not interpret `"People('foo"` or `"bar')"` as meaningful tokens. In other words, the prior behavior produced an invalid tokenization of an already non‑conformant URL. 
- The new behavior does not loosen standards: it continues to treat raw `/` as a separator, but only when outside a quoted literal. It simply avoids making the situation *worse* when a client neglected to encode the slash that is clearly intended as data. This better matches the intent of RFC 3986’s delimiter rules and keeps `%2F` vs `/` semantics intact. 

**Conclusion:** While “breaking” in a narrow sense, the previous result was not a useful contract to depend on. The new behavior is a corrective change that improves interoperability without changing the meaning of any valid OData URL.

### Additional arguments in favor

1. **Spec‑conformant by default; robust by necessity**  
   We continue to encourage clients to percent‑encode slashes inside literals (as required by OData 4.01). But servers in the wild must interoperate with clients that don’t always follow this strictly. Preserving the literal as a single segment is the most faithful interpretation of the client’s intent and avoids nonsensical segmentation. 

2. **No ambiguity introduced**  
   - We only treat `/` as a separator outside quotes.  
   - Within quotes, we preserve the raw text (including any `/`) and still honor `%27` / `%27%27` and doubled `'` escaping semantics.  
   - Balance checks ensure we error on unbalanced quotes or malformed `%` sequences as before.  
   This makes the change predictable and bounded to a well‑defined compatibility corner case. *(RFC 3986 supports component‑aware parsing; we apply it at the path level.)* 

3. **No change for encoded data**  
   Any client already sending `%2F` inside the literal is unaffected (same result before and after). 

### Risk analysis

**Who might be affected?**
- Downstream code that *explicitly relied* on the old, split behavior (e.g., custom middleware that expects two segments when encountering a raw `/` inside quotes) could observe different segment counts.  
- Security posture: By treating the entire quoted token as one segment, we’re not widening traversal or normalization issues; we still operate on the escaped path and only accept `/` as a separator outside quotes. This reduces ambiguous parses compared to the old behavior that split mid‑literal. 

**Likelihood & impact**
- Likelihood: low - the previous two‑segment shape was not a valid OData path tokenization, so relying on it was unlikely and fragile.  
- Impact: low‑to‑moderate - if such reliance exists, migration is straightforward (treat the quoted literal as a single segment and apply custom parsing inside the literal if truly needed).

Depending on how the maintainers view this change, we can:
- Ship it in a minor release - Would it be worth putting it behind a feature flag?, or  
- Defer the change to the **next major release - ODL 9** with a deprecation note now.  
- My personal recommendation is to merge as a **bug fix** that corrects an inapplicable prior outcome and improves standards alignment with OData and RFC 3986. 

### Memory & performance benchmarks

| Method         | Path                                  | Mean        | Error     | StdDev     | Median      | Ratio | RatioSD | Gen0   | Allocated | Alloc Ratio |
|--------------- |-------------------------------------- |------------:|----------:|-----------:|------------:|------:|--------:|-------:|----------:|------------:|
| Parse_Original | Customers                             | 1,508.55 ns | 37.259 ns | 106.902 ns | 1,462.45 ns |  1.00 |    0.10 | 0.1564 |    1984 B |        1.00 |
| Parse_Revised  | Customers                             |    73.09 ns |  1.766 ns |   5.039 ns |    70.40 ns |  0.05 |    0.00 | 0.0254 |     320 B |        0.16 |
|                |                                       |             |           |            |             |       |         |        |           |             |
| Parse_Original | Customers('abc')                      | 1,407.93 ns | 27.255 ns |  27.989 ns | 1,397.31 ns |  1.00 |    0.03 | 0.1640 |    2080 B |        1.00 |
| Parse_Revised  | Customers('abc')                      |    85.10 ns |  3.020 ns |   8.666 ns |    81.30 ns |  0.06 |    0.01 | 0.0286 |     360 B |        0.17 |
|                |                                       |             |           |            |             |       |         |        |           |             |
| Parse_Original | Customers('abc/pqr')                  | 1,684.34 ns | 20.034 ns |  15.641 ns | 1,687.56 ns |  1.00 |    0.01 | 0.1717 |    2168 B |        1.00 |
| Parse_Revised  | Customers('abc/pqr')                  |    99.18 ns |  3.027 ns |   8.437 ns |    95.80 ns |  0.06 |    0.01 | 0.0305 |     384 B |        0.18 |
|                |                                       |             |           |            |             |       |         |        |           |             |
| Parse_Original | Customers('abc%2Fpqr')                | 1,634.20 ns | 32.633 ns |  90.967 ns | 1,593.80 ns |  1.00 |    0.08 | 0.1736 |    2200 B |        1.00 |
| Parse_Revised  | Customers('abc%2Fpqr')                |   119.83 ns |  2.539 ns |   7.202 ns |   117.24 ns |  0.07 |    0.01 | 0.0370 |     464 B |        0.21 |
|                |                                       |             |           |            |             |       |         |        |           |             |
| Parse_Original | Customers('O''Neil')/Orders(42)/Items | 1,595.63 ns | 31.785 ns |  82.047 ns | 1,556.91 ns |  1.00 |    0.07 | 0.1907 |    2400 B |        1.00 |
| Parse_Revised  | Customers('O''Neil')/Orders(42)/Items |   153.86 ns |  3.106 ns |   8.128 ns |   151.02 ns |  0.10 |    0.01 | 0.0427 |     536 B |        0.22 |
|                |                                       |             |           |            |             |       |         |        |           |             |
| Parse_Original | Customers(1)                          | 1,441.38 ns | 35.204 ns |  99.293 ns | 1,419.87 ns |  1.00 |    0.09 | 0.1602 |    2032 B |        1.00 |
| Parse_Revised  | Customers(1)                          |    83.49 ns |  3.155 ns |   9.052 ns |    81.80 ns |  0.06 |    0.01 | 0.0267 |     336 B |        0.17 |
|                |                                       |             |           |            |             |       |         |        |           |             |
| Parse_Original | People(%27O%27%27Neil%27)/Orders(1)   | 1,824.30 ns | 36.504 ns |  94.228 ns | 1,783.86 ns |  1.00 |    0.07 | 0.1888 |    2392 B |        1.00 |
| Parse_Revised  | People(%27O%27%27Neil%27)/Orders(1)   |   166.14 ns |  3.243 ns |   7.451 ns |   162.98 ns |  0.09 |    0.01 | 0.0432 |     544 B |        0.23 |

### Tests
- Adequate tests were added to ensure maximum coverage

## Checklist (Uncheck if it is not completed)

- [x] *Test cases added*
- [x] *Build and test with one-click build and test script passed*