@xuzhg 2025-09-03
@gathogojr 
How about the following scenarios:

### /customers('abc/efg') 
  ==>Basic scenario to support, It should work.
### /customers('abc%2F/''efg')/  abc   /'%2F
  ==> Mixed the %2F and / together?
### /customers(%27abc%2F/'%27efg')/  abc'   /'%2F 
 ==> Be noted, how to distinguish the / between abc' and the last '%2f